requirements.txt
transformers
accelerate
diffusers
datasets
huggingface-hub
tokenizers
safetensors
peft
ftfy
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
albumentations
opencv-python-headless
imageio
Pillow
webdataset
img2dataset
blobfile
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
pyarrow
PyYAML
rich
fire
psutil
flask
flask-cors
requests